Add casting to short for R-R sample values. Made as a workaround for an issue in Zephyr's Android library that prevents negative value of RtoR to be correctly represented.

+++ app/src/main/java/eu/fistar/sdcs/pa/da/zephyrbh/ZephyrBHConnectedListener.java
@@ -372,7 +372,7 @@ public class ZephyrBHConnectedListener extends ConnectListenerImpl {
 
         // Convert short values to String
         for (int i = 0; i < samples.length; i++) {
-            strSamples[i] = Integer.toString(samples[i]);
+            strSamples[i] = Short.toString((short) samples[i]);
         }
 
         // Create the Observation object
